# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"AuthorDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CardVisibilityEnum": {
				"enum": [
					"private",
					"public",
					"unlisted"
				],
				"type": "string"
			},
			"CardDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"mediaUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"visibility": {
						"$ref": "#/components/schemas/CardVisibilityEnum"
					},
					"author": {
						"$ref": "#/components/schemas/AuthorDto"
					},
					"isLiked": {
						"type": "boolean"
					},
					"isFavorited": {
						"type": "boolean"
					},
					"likes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"favorites": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"comments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"category": {
						"type": "string"
					},
					"gradient": {
						"type": "string"
					},
					"allowComments": {
						"type": "boolean"
					},
					"allowDownloads": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"mediaUrls",
					"visibility",
					"author",
					"likes",
					"favorites",
					"comments",
					"tags",
					"category",
					"gradient",
					"allowComments",
					"allowDownloads",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponseDto_CardDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/CardDto"
						},
						"type": "array"
					},
					"nextCursor": {
						"type": "string"
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserRefDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string",
						"nullable": true
					},
					"bio": {
						"type": "string"
					},
					"cards": {
						"items": {
							"$ref": "#/components/schemas/CardDto"
						},
						"type": "array"
					},
					"favorites": {
						"items": {
							"$ref": "#/components/schemas/CardDto"
						},
						"type": "array"
					},
					"likes": {
						"items": {
							"$ref": "#/components/schemas/CardDto"
						},
						"type": "array"
					},
					"isFollowing": {
						"type": "boolean"
					},
					"following": {
						"items": {
							"$ref": "#/components/schemas/UserRefDto"
						},
						"type": "array"
					},
					"followers": {
						"items": {
							"$ref": "#/components/schemas/UserRefDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"fullName",
					"username",
					"profilePicture",
					"bio",
					"cards",
					"favorites",
					"likes",
					"isFollowing",
					"following",
					"followers"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "cardshare-api",
		"version": "1.0.0",
		"description": "CardShare API with Express.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"$ref": "#/components/schemas/UserDto"
										}
									},
									"required": [
										"user"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "User not authenticated"
					},
					"404": {
						"description": "User not found"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/username/{username}": {
			"get": {
				"operationId": "GetUserByUsername",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"$ref": "#/components/schemas/UserDto"
										}
									},
									"required": [
										"user"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "User not authenticated"
					},
					"404": {
						"description": "User not found"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{id}/follow": {
			"post": {
				"operationId": "ToggleFollowUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"following": {
											"type": "boolean"
										}
									},
									"required": [
										"following"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "User not authenticated"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/me/cards": {
			"get": {
				"operationId": "GetMyCards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponseDto_CardDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"latest",
								"most-liked"
							]
						}
					}
				]
			}
		},
		"/users/me/liked": {
			"get": {
				"operationId": "GetMyLikedCards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponseDto_CardDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"latest",
								"most-liked"
							]
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}